% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{col2plotlyrgb}
\alias{col2plotlyrgb}
\title{Color to RGB Text}
\usage{
col2plotlyrgb(col)
}
\arguments{
\item{col}{vector of any of the three kinds of R color specifications,
i.e., either a color name (as listed by colors()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb),
or a positive integer i meaning palette()[i].}
}
\value{
A character of the form "rgb(value1,value1,value3)"
}
\description{
Plotly takes colors in this format "rgb(255, 0, 0)"
}
\examples{
\dontrun{
col2rgb("peachpuff")
col2plotlyrgb("peachpuff")
}
}
\seealso{
\link{col2rgb}
}
