% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\alias{ept.data.frame}
\alias{ept.default}
\title{Empirical Percentile Transformation}
\usage{
ept(x, ...)
}
\arguments{
\item{x}{a vector or a data.frame.}

\item{...}{Currently ignored.}
}
\value{
A vector (or data.frame) after \link{ecdf} was used on that vector.
If x is a \link{data.frame} then only the numeric variables are transformed.
}
\description{
An Empirical Percentile Transformation (ept) is similar to taking the rank
of a variable. The difference is that it is simpler to compare and interpret
the transformed variables.

This is helpful for comparing several variables in a heatmap (e.g.: \link{heatmaply}).
}
\examples{
\dontrun{
x <- mtcars
x <- data.frame(x)
x$am <- factor(x$am)
x$vs <- factor(x$vs)
heatmaply(ept(x))

}
}

