% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Creates a ggplot2 heatmap}
\usage{
ggheatmap(x, ...)
}
\arguments{
\item{x}{can either be a heatmapr object, or a numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}

\item{...}{other parameters passed to \link{heatmapr} (currently, various parameters may be ignored.}
}
\description{
An object of class heatmapr includes all the needed information
for producing a heatmap. The goal is to seperate the pre-processing of the
heatmap elements from the graphical rendaring of the object, which could be done
}
\examples{
\dontrun{

library(heatmaply)
x <- heatmapr(iris[,-5], scale = "column", colors = "Blues")
ggheatmap(x)


}
}
