% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{Creates a ggplot2 heatmap}
\usage{
ggplot_heatmap(xx, row_text_angle = 0, column_text_angle = 45,
  scale_fill_gradient_fun = scale_fill_gradientn(colors = viridis(n = 256,
  alpha = 1, begin = 0, end = 1, option = "viridis"), na.value = "grey50",
  limits = NULL), grid_color = NA, grid_size = 0.1, key.title = NULL,
  layers, row_dend_left = FALSE, ...)
}
\arguments{
\item{...}{other parameters passed to \link{heatmapr} (currently, various parameters may be ignored.}

\item{x}{can either be a heatmapr object, or a numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}
}
\description{
An object of class heatmapr includes all the needed information
for producing a heatmap. The goal is to seperate the pre-processing of the
heatmap elements from the graphical rendaring of the object, which could be done
}
\examples{
\dontrun{

library(heatmaply)
x <- heatmapr(iris[,-5], scale = "column", colors = "Blues")
ggheatmap(x)


}
}

