% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_side_color_plot}
\alias{ggplot_side_color_plot}
\title{Side color plots for heatmaps}
\usage{
ggplot_side_color_plot(df, palette = NULL, scale_title = paste(type,
  "side colors"), type = c("column", "row"), text_angle = if (type ==
  "column") 0 else 90, is_colors = FALSE, label_name = type)
}
\arguments{
\item{df}{A "molten" data.frame as produced by (eg) reshape2::melt}

\item{palette}{A function which can return colors to be used in the sidebar
plot}

\item{scale_title}{Title of the color scale. Not currently used.}

\item{type}{Horizontal or vertical plot? Valid values are "column" and "row"}

\item{text_angle}{the angle of the text of the rows/columns.}

\item{is_colors}{Use if the values in df are valid colours and should not be mapped
to a color scheme, and instead should be plotted directly.}

\item{label_name}{Name for the mouseover label, usually "row" or "column"}
}
\value{
A ggplot geom_tile object
}
\description{
Important for creating annotation.
}
