% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapr.R
\name{is.heatmapr}
\alias{is.heatmapr}
\title{Is the object of class heatmapr}
\usage{
is.heatmapr(x)
}
\arguments{
\item{x}{an object.}
}
\value{
logical - is the object of class heatmapr.
}
\description{
Is the object of class heatmapr.
}

