% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaply.R
\name{is.plotly}
\alias{is.plotly}
\title{Checks if an object is of class plotly or not.}
\usage{
is.plotly(x)
}
\arguments{
\item{x}{an object to check}
}
\value{
TRUE if the object ingerits "plotly" as a class.
}
\description{
Helpful for the plot_method in link{heatmaply}.
}
