% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_colorscale}
\alias{make_colorscale}
\title{Create a plotly colorscale from a list of colors in any format. 
Probably not needed currently}
\usage{
make_colorscale(colors)
}
\description{
Create a plotly colorscale from a list of colors in any format. 
Probably not needed currently
}
