% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{ggplot heatmap equivalent to heatmaply}
\usage{
ggheatmap(
  ...,
  widths = NULL,
  heights = NULL,
  row_dend_left = FALSE,
  hide_colorbar = FALSE
)
}
\arguments{
\item{...}{Passed to \link{heatmaply}}

\item{widths, heights}{Relative widths and heights of plots.}

\item{row_dend_left}{Logical argument controlling whether the row
dendrogram is placed on the left of the plot.}

\item{hide_colorbar}{Logical argument controlling whether the color bar (i.e.:
the legend) is hidden.}
}
\description{
This function produces a ggplot analogue of heatmaply figures
using \link[egg]{ggarrange}. This function may not always support the same
set of features as , and exporting the heatmaply object with, for example,
\link[plotly]{orca} or \code{heatmaply(mtcars, file = "foo.png")}.
}
\examples{
ggheatmap(mtcars)
}
